pyproject.toml
[project]
name = "rss-agg"
version = "0.1.0"
description = "Aggregate, de-duplicate and republish RSS feeds"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.56",
    "defusedxml>=0.7.1",
    "flask[async]>=3.1.2",
    "gunicorn>=25.0.1",
    "httpx[http2]>=0.28.1",
    "python-json-logger>=4.0.0",
    "wireup>=2.6.0",
    "yarl>=1.22.0",
]

[dependency-groups]
dev = [
    "brunns-matchers>=2.10.0",
    "mbtest>=2.14.0",
    "mockito>=1.5.5",
    "pyfakefs>=6.0.0",
    "pyhamcrest>=2.1.0",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-docker>=3.2.5",
    "pytest-mockito>=0.0.6",
    "respx>=0.22.0",
    "ruff>=0.14.14",
    "scripttest>=2.0.post1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rss_agg"]

[project.scripts]
cli = "rss_agg.cli:main"

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "ISC001", "D", "PGH003", "FBT001", "FBT002"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S105", "S106", "S108", "PLR2004", "E501", "ANN", "INP", "S314"]
"lambda_function.py" = ["ANN401"]
"tests/integration/test_cli.py" = ["PTH109"]

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.coverage.run]
omit = ["rss_agg/cli.py", "rss_agg/web.py", "rss_agg/logging_config.py"]

[tool.pyright]
include = ["rss_agg/"]
pythonVersion = "3.14"
